+++ src/data/product.ts
@@ -145,7 +145,7 @@ export const productData = {
     }
   },
   "release": {
-    "version": "3.7.2"
+    "version": "3.7.3"
   },
   "extension": {
     "itemName": "hearth-code.hearth-theme",